Update review models and schemas to use float ratings; enhance todo with detailed rating descriptions and user story

+++ backend/FacultyManagement/models/review.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Date, func
+from sqlalchemy import Column, Integer, Float, String, ForeignKey, Date, func
 from sqlalchemy.orm import relationship
 from database.database import Base
 
@@ -8,9 +8,9 @@ class Review(Base):
     id = Column(Integer, primary_key=True, index=True)
     course_id = Column(Integer, ForeignKey("courses.id"), nullable=False)
     student_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
-    rating_clarity = Column(Integer, nullable=False)
-    rating_feasibility = Column(Integer, nullable=False)
-    rating_availability = Column(Integer, nullable=False)
+    rating_clarity = Column(Float, nullable=False)
+    rating_feasibility = Column(Float, nullable=False)
+    rating_availability = Column(Float, nullable=False)
     comment = Column(String, nullable=True)
     created_at = Column(Date, server_default=func.current_date())
